pyproject.toml
[project]
name = "modal-server"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.1.1",
    "python-dotenv>=1.0.1",
    "modal>=0.67",
]
[[project.authors]]
name = "Server Modal"
email = "team@example.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
modal-server = "modal_server:main"
